+++ assets/pages/new-contributions-landing/components/ContributionSubmit.jsx
@@ -9,7 +9,7 @@ import { connect } from 'react-redux';
 import { getFrequency, type ContributionType, type PaymentMethod } from 'helpers/contributions';
 import { getPaymentDescription } from 'helpers/checkouts';
 import { type IsoCurrency, currencies, spokenCurrencies } from 'helpers/internationalisation/currency';
-import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/newPaymentFlow/readerRevenueApis';
+import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
 import type { SelectedAmounts } from 'helpers/contributions';
 import { hiddenIf } from 'helpers/utilities';
 import { type State } from '../contributionsLandingReducer';
