+++ assets/pages/new-contributions-landing/components/SetPasswordForm.jsx
@@ -7,7 +7,7 @@ import { connect } from 'react-redux';
 import { type Dispatch } from 'redux';
 
 import { classNameWithModifiers } from 'helpers/utilities';
-import { setPasswordGuest } from 'helpers/paymentIntegrations/newPaymentFlow/readerRevenueApis';
+import { setPasswordGuest } from 'helpers/paymentIntegrations/readerRevenueApis';
 import type { Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
 import SvgPasswordKey from 'components/svgs/passwordKey';
 import SvgEnvelope from 'components/svgs/envelope';
