+++ assets/pages/new-contributions-landing/components/StripePaymentRequestButton/StripePaymentRequestButtonContainer.jsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import { StripeProvider, Elements } from 'react-stripe-elements';
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
-import { getStripeKey } from 'helpers/paymentIntegrations/newPaymentFlow/stripeCheckout';
+import { getStripeKey } from 'helpers/paymentIntegrations/stripeCheckout';
 import type { ContributionType, OtherAmounts, SelectedAmounts } from 'helpers/contributions';
 import { getAmount } from 'helpers/contributions';
 import { isInStripePaymentRequestAllowedCountries } from 'helpers/internationalisation/country';
